Re-write an arcana-laden sentence, fix a funny bug
I re-wrote a sentence that contained the following **arcane** phrases: "Win16 applications", "OS/2 applications", and "MS-DOS applications". Understanding these phrases was in no way essential to understanding this document.

I fixed a rather amusing bug too. The following part of the sentence was rendered in italics:

> `*.bat), command files (*`

The reason is the asterisk characters. These characters are part of the Markdown syntax. Of course, using them in this context is semantically wrong too; this cmdlet does not support wildcards.