Address Misleading Error
This is a proposal to add a notice addressing a misleading error in shell for single instance deployments. See below for more information.

Symptoms
Assume that you use the inbox Windows Internal Database (WID) in Windows Server. If you disable Transport Layer Security (TLS) 1.0 when you configure security settings, you experience the following issues:
•	The Remote Desktop service (RDS) may fail.
•	An existing RDS deployment that uses Remote Desktop Connection Broker and WID may fail.
•	The Remote Desktop Management service (RDMS) doesn't start.
•	You receive the following error message when you try to start the RDMS:
The Remote Desktop Management service on Local Computer started and then stopped. Some services stop automatically if they are not in use by other services or programs.
•	The Remote Desktop Connection Broker role can't be installed.
•       Server is pending a reboot
•       RemoteDesktop Roles is not installed (Even if it is installed)

Cause
This is expected behavior. This is because of the current dependencies between RDS and Windows Internal Database. RDMS and Connection Broker depend on TLS 1.0 to authenticate with the database. WID doesn't currently support TLS 1.2. So, disabling TLS 1.0 breaks this communication.

Resolution
To fix this issue, use one of the following methods:
•	Set up RDS without Connection Broker for a single-server installation.
•	Don't disable TLS 1.0 on a single Connection Broker deployment.
•	Configure a high availability Connection Broker deployment that uses dedicated SQL Server.

Article Link:
https://docs.microsoft.com/en-us/troubleshoot/windows-server/remote/rds-connection-broker-or-rdms-fails-caused-by-disabled-tls-10