Update Set-MpPreference.md
Changed documentation for -

CheckForSignaturesBeforeRunningScan (Xudong confirmed that we dont check for signatures before running a scan from the command line as manual scan takes many shapes and forms and can happen on a single file/resource. It is supposed to be fast. Always checking for update on those would be problematic.

DisableEmailScanning - This feature is not applicable to modern email clients.

DisableArchiveScanning - Based on the code we area always scanning archives during directed scans