Added the location of the export zone file
Added the location to where the export zone file can be found (C:\Windows\System32\DNS) and that the location is on the DNS server.  Also added a note that -FileName is a filename and not a filepath.  This should clarify how users can use this command and where they can locate the output.