Update Update-IscsiTargetPortal.md
Based on requested change "it would look better if you enclose the commands and output"... The command and output were _already_ enclosed in a "```" block, starting with lines way above the one line that was changed for this patch. But the existing lines were enclosed in a single "```". So, not being clear about what was being asked to be changed, since only the last line within that already existing block was changed... simply added the "Output" tag between the commands and output - but again, that's _way_ above the line that was changed for this patch "patch-2".