(MAINT) Add cSpell to the repository
This change introduces cSpell to the repository for managing spelling
across the files. In this initial draft, it adds the upstream dictionary
from MicrosoftDocs/PowerShell-Docs and scaffolds a nearly-empty
dictionary for the Windows modules to make it easier to define terms in
the future.

It also removes the directive in gitignore so that the .vscode folder
will be tracked.