(MAINT) Add issue templates
This change introduces GitHub issue templates and the issue picker to
the project, enabling an improved UX for people filing issues against
the repository. It also includes the issue template for Hacktoberfest
2022, to make participation easier.

The configuration for the article issue template is minimal, lacking
version or module selection, but those fields can be added in the future
if required.