Update BackupToAAD-BitLockerKeyProtector.md
Assuming that the RecoveryPassword protector type is always at index 1 in the array can lead to inconsistent results—I experienced this just yesterday. It's better to explicitly select the correct protector type rather than relying on it always being at index 1. On one system I worked on, it was actually at index 0, which is why I'm making this suggestion.